Explicit logging of analyzer output
The scan-build-py analyzer wrapper now returns the error code of the
analyzer invocation, rather than the compiler invocation, in order to
indicate if the analyzer succeeded or failed.

The xtu-analyze script drops the output of the scan-build invocation
into a file, and also dumps it to terminal if the verbose switch is
passed. The file is in a directory called "passes" or "fails" depending
on the return code of the scan-build-py invocation.

These changes are to help diagnose bugs with the AST Importer, as it is
otherwise difficult to tell which files caused the analyzer to break and
why.